i18n_embed

Module fluent

This module contains the types and functions to interact with the fluent localization system.

Most important is the FluentLanguageLoader.

⚠️ This module requires the following crate features to be activated: fluent-system.

Macros

fluent_language_loader

A procedural macro to create a new FluentLanguageLoader using the current crate’s i18n.toml configuration, and domain.

Structs

FluentLanguageLoader

LanguageLoader implementation for the fluent localization system. Also provides methods to access localizations which have been loaded.

Enums

NegotiationStrategy