i18n_embed

Struct DesktopLanguageRequester

pub struct DesktopLanguageRequester<'a> { /* private fields */ }

A LanguageRequester for the desktop platform, supporting windows, linux and mac. It uses sys-locale to select the language based on the system selected language.

⚠️ This API requires the following crate features to be activated: desktop-requester.

Implementations

impl DesktopLanguageRequester<'_>

pub fn new() -> Self

Create a new DesktopLanguageRequester.

pub fn requested_languages() -> Vec<LanguageIdentifier>

The languages being requested by the operating system/environment according to the [sys-locale] crate’s implementation.