i18n_embed

Trait I18nAssets

pub trait I18nAssets {
    // Required methods
    fn get_files(&self, file_path: &str) -> Vec<Cow<'_, [u8]>>;
    fn filenames_iter(&self) -> Box<dyn Iterator<Item = String> + '_>;

    // Provided method
    fn subscribe_changed(
        &self,
        changed: Arc<dyn Fn() + Send + Sync + 'static>,
    ) -> Result<Box<dyn Watcher + Send + Sync + 'static>, I18nEmbedError> { ... }
}

A trait to handle the retrieval of localization assets.

Required Methods

fn get_files(&self, file_path: &str) -> Vec<Cow<'_, [u8]>>

Get localization asset files that correspond to the specified file_path. Returns an empty Vec if the asset does not exist, or unable to obtain the asset due to a non-critical error.

fn filenames_iter(&self) -> Box<dyn Iterator<Item = String> + '_>

Get an iterator over the file paths of the localization assets. There may be duplicates where multiple files exist for the same file path.

Provided Methods

fn subscribe_changed( &self, changed: Arc<dyn Fn() + Send + Sync + 'static>, ) -> Result<Box<dyn Watcher + Send + Sync + 'static>, I18nEmbedError>

A method to allow users of this trait to subscribe to change events, and reload assets when they have changed. The subscription will be cancelled when the returned Watcher is dropped.

NOTE: The implementation of this method is optional, don’t rely on it functioning for all implementations.