i18n_embed

Trait Localizer

pub trait Localizer {
    // Required methods
    fn language_loader(&self) -> &dyn LanguageLoader;
    fn i18n_assets(&self) -> &dyn I18nAssets;

    // Provided methods
    fn available_languages(
        &self,
    ) -> Result<Vec<LanguageIdentifier>, I18nEmbedError> { ... }
    fn select(
        &self,
        requested_languages: &[LanguageIdentifier],
    ) -> Result<Vec<LanguageIdentifier>, I18nEmbedError> { ... }
}

This trait provides dynamic access to an LanguageLoader and an I18nAssets, which are used together to localize a library/crate on demand.

Required Methods

fn language_loader(&self) -> &dyn LanguageLoader

The LanguageLoader used by this localizer.

fn i18n_assets(&self) -> &dyn I18nAssets

The source of localization assets used by this localizer

Provided Methods

fn available_languages(&self) -> Result<Vec<LanguageIdentifier>, I18nEmbedError>

The available languages that can be selected by this localizer.

fn select( &self, requested_languages: &[LanguageIdentifier], ) -> Result<Vec<LanguageIdentifier>, I18nEmbedError>

Automatically the language currently requested by the system by the the LanguageRequester), and load it using the provided LanguageLoader.