i18n_embed

Trait Watcher

pub trait Watcher { }

An entity that watches for changes to localization resources.

NOTE: Currently we rely in the implicit Drop implementation to remove file system watches, in the future there may be new methods added to this trait.