i18n_embed

Trait LanguageRequester

pub trait LanguageRequester<'a> {
    // Required methods
    fn add_listener(&mut self, listener: Weak<dyn Localizer>);
    fn add_listener_ref(&mut self, listener: &'a dyn Localizer);
    fn poll(&mut self) -> Result<(), I18nEmbedError>;
    fn set_language_override(
        &mut self,
        language_override: Option<LanguageIdentifier>,
    ) -> Result<(), I18nEmbedError>;
    fn requested_languages(&self) -> Vec<LanguageIdentifier>;
    fn available_languages(
        &self,
    ) -> Result<Vec<LanguageIdentifier>, I18nEmbedError>;
    fn current_languages(&self) -> HashMap<String, LanguageIdentifier>;
}

A trait used by I18nAssets to ascertain which languages are being requested.

Required Methods

fn add_listener(&mut self, listener: Weak<dyn Localizer>)

Add a listener to this LanguageRequester. When the system reports that the currently requested languages has changed, each listener will have its Localizer#select() method called. Weak is used so that when the Arc that it references is dropped, the listener will also be removed next time this requester is polled/updates.

If you haven’t already selected a language for the localizer you are adding here, you may want to manually call #poll() after adding the listener/s.

fn add_listener_ref(&mut self, listener: &'a dyn Localizer)

Add a listener to this LanguageRequester. When the system reports that the currently requested languages has changed, each listener will have its Localizer#select() method called. As opposed to LanguageRequester::add_listener(), this listener will not be removed.

If you haven’t already selected a language for the localizer you are adding here, you may want to manually call #poll() after adding the listener/s.

fn poll(&mut self) -> Result<(), I18nEmbedError>

Poll the system’s currently selected language, and call Localizer#select() on each of the listeners.

NOTE: Support for this across systems currently varies, it may not change when the system requested language changes during runtime without restarting your application. In the future some platforms may also gain support for automatic triggering when the requested display language changes.

fn set_language_override( &mut self, language_override: Option<LanguageIdentifier>, ) -> Result<(), I18nEmbedError>

Override the languages fed to the Localizer listeners during a #poll(). Set this as None to disable the override.

fn requested_languages(&self) -> Vec<LanguageIdentifier>

The currently requested languages.

fn available_languages(&self) -> Result<Vec<LanguageIdentifier>, I18nEmbedError>

The languages reported to be available in the listener Localizers.

fn current_languages(&self) -> HashMap<String, LanguageIdentifier>

The languages currently loaded, keyed by the LanguageLoader::domain().